requirements.txt
requests>=2.28.0
flask>=3.0.0
fpdf2>=2.7.0
selenium>=4.15.0
